split NewSongWizard into separate components for artist, album, song
